Added experimental support for Microsoft's D3D9On12 mapping layer, which maps graphics commands from D3D9 to D3D12. (https://github.com/microsoft/D3D9On12)
Unfortunately, it's only able to display the backbuffer in exclusive fullscreen mode. Otherwise, the window will be blank.

Seeing how there are no longer any functioning D3D9 debug tools (that I could find), this is the only way to do GPU captures for debugging. RenderDoc crashes on startup, but Pix for Windows works. However, it does not retain markers, nor does it generate shader debug information, so it's still not an ideal solution.

Renamed all mentions of DX9 to the more accurate D3D9, including folder and file names.