Added CPU timings to the profiler.
Added a separate UI color buffer and swapped the order of the 'copy to backbuffer' and 'UI' passes. If we end up waiting on an available backbuffer due to vsync, there's no point in delaying UI rendering.

Removed the half pixel offset patch in the UI's vertex shader, since it caused issues with text rendering at low resolutions.